+++ SandboxiePlus/SandMan/Windows/NewBoxWindow.cpp
@@ -69,9 +69,12 @@ void CNewBoxWindow::OnBoxTypChanged()
 void CNewBoxWindow::CreateBox()
 {
 	m_Name = ui.txtName->text();
-	m_Name.replace(" ", "_");
-
 	int BoxType = ui.cmbBoxType->currentData().toInt();
+
+	if (!theGUI->GetBoxView()->TestNameAndWarn(m_Name))
+		return;
+
+	m_Name.replace(" ", "_");
 
 	SB_STATUS Status = theAPI->CreateBox(m_Name, true);