+++ CHANGELOG.md
@@ -12,10 +12,23 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.0.19 / 5.55.19] - 2022-04-??
 
+### Added
+- added drag and drop support for groups [#1775](https://github.com/sandboxie-plus/Sandboxie/issues/1775)
+- added del key support to the box view for all entry types [#1779](https://github.com/sandboxie-plus/Sandboxie/issues/1779)
+- added warnign when trying to run explorer.exe in a box with OpenCOM [#1716](https://github.com/sandboxie-plus/Sandboxie/issues/1716)
+
 ### Fixed
-- fixed crash issue in the sandman ui
+- fixed crash issue in the sandman ui [#1772](https://github.com/sandboxie-plus/Sandboxie/issues/1772)
 - fixed issue some installers when EnableObjectFiltering is enabled [#1795](https://github.com/sandboxie-plus/Sandboxie/issues/1795)
 - fixed to allow NtCreateSymbolicLinkObject to be used safely in the sandbox
+- added workaround for a vivaldi hooking issue [1783](https://github.com/sandboxie-plus/Sandboxie/issues/1783)
+-- Note: its a very provisional fix hence it can be disabled with UseVivaldiWorkaround=n
+- fixed registry issue with snapshots [#1782](https://github.com/sandboxie-plus/Sandboxie/issues/1782)
+- fixed issue with box grouping [#1778](https://github.com/sandboxie-plus/Sandboxie/issues/1778) [#1777](https://github.com/sandboxie-plus/Sandboxie/issues/1777) [#1776](https://github.com/sandboxie-plus/Sandboxie/issues/1776) 
+- fixed more issue with box grouping [#1698](https://github.com/sandboxie-plus/Sandboxie/issues/1698) [#1697](https://github.com/sandboxie-plus/Sandboxie/issues/1697)
+- fixed issues with snadshot ui [#1696](https://github.com/sandboxie-plus/Sandboxie/issues/1696) [#1695](https://github.com/sandboxie-plus/Sandboxie/issues/1695)
+
+